Extract create/update/delete helpers to api/mutation
Move the biggest remaining slice of api/__init__.py into a focused
module: the three top-level route orchestrators (create_items,
update_items, delete_items) plus their patch/put aliases, the
low-level building blocks (create_object, update_object_with_data,
deactivate_object, create_analysisrequest, find_target_container,
validate_object) and the two permission checks (is_creation_allowed,
is_update_allowed).

All names remain importable from senaite.jsonapi.api via explicit
re-exports at the bottom of __init__.py. v1/routes/content.action
looks up 'api.create_items' / 'api.update_items' / 'api.delete_items'
via getattr and continues to work unchanged.

Semantic tightening while moving:

- Every 'fail(401, ...)' inside the moved code becomes a proper
  ForbiddenError. HTTP 401 means 'log in and try again'; the fail
  sites (denied by permission gate, denied by adapter, container
  disallows type, ...) all mean 'you are authenticated but may not
  do this' -- which is 403. The two affected doctests
  (create.rst, update.rst) update their expected status from
  401 to 403 to match.

- 'fail(400, ...)' becomes BadRequestError, 'fail(404, ...)' becomes
  NotFoundError. Backward compat is preserved through the shared
  APIError base.

Drop now-unused imports from __init__.py: copy, transaction,
AccessControl.Unauthorized, create_ar, ICreate, IUpdate, IInfo,
getAdapters, zope.deprecation.deprecate.
